requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Security==4.1.5
Flask-Admin==1.6.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
cloudinary==1.37.0
Werkzeug==3.0.1
email-validator==2.1.0.post1
gunicorn==21.2.0